refactor: reorganize file structure and enhance functionality
- Updated import paths for CSS and HTML generation utilities to reflect new directory structure.
- Removed the obsolete generateHexagons.drawio file.
- Introduced new generator functions for CSS and HTML, including tests for their functionality.
- Added highlightText utility for formatting HTML and CSS strings.
