+++ src/app/renderHexagonSection.ts
@@ -1,4 +1,4 @@
-import { generateHexagons } from "../lib/generateHexagons";
+import { generateHexagons } from "../lib/generators/generateHexagons";
 import type { AppElements } from "./appElements";
 import type { AppState } from "./readAppState";
