+++ src/lib/formatters/highlightText.spec.ts
