+++ src/lib/formatters/highlightText.ts
